Corrige exclusão de exemplar emprestado Não permite excluir exemplar caso ele esteja emprestado.
Corrige botão de devolução
Antes o formulário de devolução de material (emprestimos.edit) constava como método patch, que não excluía o empréstimo.

Fix #45

+++ app/Http/Controllers/InstanceController.php
@@ -82,6 +82,9 @@ class InstanceController extends Controller
     public function destroy(Instance $instance)
     {
         $this->authorize('admin');
+        if ($instance->emprestimos->isNotEmpty()){
+            return redirect('/records')->with('alert-danger', 'Exemplar está emprestado. Por favor, desfaça o empréstimo do exemplar antes!');
+        }
         $instance->delete();
         return redirect('/records');
     }
